Fix citations for submission
After submission, I was informed by JGR that I had a citation that was
in press or in prep without a DOI, and that needed to be fixed before my
paper could move forward. This commit updates that citation (the Webb
et al. CFMIP contributions to CMIP6 manuscript), as well as fixes a few
more citations that still had abbreviations for journal names.